// src/services/api/generated/inventory-service/model/createCinemaRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Cinema Booking System API
 * API documentation for Cinema Booking System
 * OpenAPI spec version: 1.0.0
 */

export interface CreateCinemaRequest {
  /**
     * @minLength 0
     * @maxLength 150
     */
  name: string;
  /**
     * @minLength 0
     * @maxLength 500
     */
  address: string;
  /**
     * @minLength 0
     * @maxLength 100
     */
  city: string;
}